// src/app/types/user.ts
// types.ts
export type User = {
    id_user: number;
    name: string;
    surname: string;
    mail: string;
    password: string;
    role: string;
    phone: string;
    street: string;
    city: string;
    postal_code: number;
    status: string;
    code_referral: string;
    id_sponsor: number;
  };